swap movements
Implemented the swap operations for the stack, including the functions `sa`, `sb`, and `ss`, which operate at a higher level by calling the lower-level `swap` function to interchange the first two elements of a stack. Additionally, added the `ft_lst_orchestor` function to initialize the list from an entry. These changes establish a clear hierarchy of function calls, with the movement orchestrator capable of invoking sub-level functions that perform the actual swap operations.

 - sw_moves.c
	|
	-> swap.c