rotate movements
Implemented rotate operations for the stack, including the functions `ra`, `rb`, and `rr`, which operate at a higher level by calling the lower-level `ft_rotate` function to rotate the elements of a stack. Additionally, added the `ft_rotates.c` file to define the `ft_rotate` function, which performs the actual rotation of the stack elements. These changes establish a clear hierarchy of function calls, with the movement orchestrator capable of invoking sub-level functions that perform the actual rotate operations.

 - rt_moves.c -> ft_rotate.c